Delindex
added deleteindex plus extra commands

+++ modules/eightball.py
@@ -4,7 +4,7 @@ def eightball(phenny, input):
     foo = open("8ball.txt")
     phrases = (foo.read()).split("^~`")
     phenny.say(phrases[randrange(len(phrases))])
-eightball.commands = ['8']
+eightball.commands = ['8', '!8', '8ball']
 
 def addresponse(phenny, input):
     foo = open("8ball.txt")
@@ -18,7 +18,7 @@ def addresponse(phenny, input):
         foo.write(newphrases)
         phenny.say(str(newphrases.split("^~`")))
     
-addresponse.commands = ['8add']
+addresponse.commands = ['8add', '8balladd']
 
 def removeresponse(phenny, input):
     foo = open("8ball.txt")
@@ -35,10 +35,34 @@ def removeresponse(phenny, input):
         phenny.say(str(nowphrases.split("^~`")))
 removeresponse.commands = ["8del"]
 
+def removeindex(phenny,input):
+    foo = open("8ball.txt")
+    stuff= ((input.group()).split())
+    try:
+        index = int(stuff[1])
+        cont=True
+    except:
+        phenny.say("Not an integer")
+        cont=False
+    phrases = (foo.read()).split("^~`")
+    if cont==True:
+        if len(phrases)<(index-1): phenny.say("Index doesn't exist")
+        else:
+            del phrases[index] 
+            nowphrases = "^~`".join(phrases)
+            foo = open("8ball.txt", 'w')
+            foo.write(nowphrases)
+            phenny.say(str(nowphrases.split("^~`")))
+removeindex.commands = ["8delindex"]
+    
+
 def balldefault(phenny, input):
-    foo = open("8ball.txt", 'w')
-    foo.write("Yes^~`No^~`Maybe^~`Ask Again Later^~`Isn't it obvious?")
-    showresponses(phenny,input)
+    if input.admin==False:
+        phenny.say("You need to be an admin to run this command")
+    else:
+        foo = open("8ball.txt", 'w')
+        foo.write("Yes^~`No^~`Maybe^~`Ask Again Later^~`Isn't it obvious?")
+        showresponses(phenny,input)
 balldefault.commands = ["8default"]
 
 def showresponses(phenny,input):
@@ -48,6 +72,6 @@ def showresponses(phenny,input):
 showresponses.commands = ["8show"]
 
 def ballhelp(phenny,input):
-    phenny.say ("Type .8 to get your answer, .8add ANSWER to add a result, .8del ANSWER to delete an existing answer, .8show to show all possible answers, and .8default to reset the answers. Have fun!")
+    phenny.say ("Type .8 to get your answer, .8add ANSWER to add a result, .8del ANSWER to delete an existing answer, .8delindex INTEGER to remove that indexed answer (start counting at 0), .8show to show all possible answers, and .8default to reset the answers. Have fun!")
 
 ballhelp.commands = ['8help']
